fine_tuning/requirements.txt
# requirements.txt

# Data Generation
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Training
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
datasets>=2.14.0
trl>=0.7.0
peft>=0.6.0

# Unsloth (fastest training)
git+https://github.com/unslothai/unsloth.git

# Optional: Constrained decoding
outlines>=0.0.34
pydantic>=2.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0